+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Shrubs, a system for spawning smaller tree-like plants into the world.
 - Waterfalls
 - Sailing boat (currently requires spawning in)
+- Added a filter search function for crafting menu, use "input:_____" to search for recipe inputs
 - Added catalan (Catalonia) language translation
 
 ### Changed